extractUserData 메소드 이동 RegistrationHandler 클라스를 지우고 안에 있었던 메소드를 UerHandler 로 옯겨주었습니다.
